+++ tests/Unit/Rules/ListOfRuleTest.php
@@ -16,7 +16,7 @@ use Orisai\ObjectMapper\Rules\StringRule;
 use Orisai\ObjectMapper\Types\GenericArrayType;
 use Orisai\ObjectMapper\Types\SimpleValueType;
 use Tests\Orisai\ObjectMapper\Doubles\Rules\AlwaysInvalidRule;
-use Tests\Orisai\ObjectMapper\Doubles\Rules\EfficientTestRule;
+use Tests\Orisai\ObjectMapper\Doubles\Rules\PhasedTestRule;
 use Tests\Orisai\ObjectMapper\Toolkit\ProcessingTestCase;
 
 final class ListOfRuleTest extends ProcessingTestCase
@@ -93,7 +93,7 @@ final class ListOfRuleTest extends ProcessingTestCase
 
 	public function testProcessMultiStepCallOrder(): void
 	{
-		$rule = new EfficientTestRule();
+		$rule = new PhasedTestRule();
 		$this->ruleManager->addRule($rule);
 
 		$value = ['foo', 'bar', 'baz', 123];
@@ -101,7 +101,7 @@ final class ListOfRuleTest extends ProcessingTestCase
 		$processed = $this->rule->processValue(
 			$value,
 			new MultiValueArgs(
-				new RuleRuntimeMeta(EfficientTestRule::class, new EmptyArgs()),
+				new RuleRuntimeMeta(PhasedTestRule::class, new EmptyArgs()),
 				null,
 				null,
 				false,
@@ -155,7 +155,7 @@ final class ListOfRuleTest extends ProcessingTestCase
 
 	public function testProcessMultiStepCallOrderWithErrors(): void
 	{
-		$rule = new EfficientTestRule();
+		$rule = new PhasedTestRule();
 		$this->ruleManager->addRule($rule);
 
 		$value = [$rule::Fail1, 42 => $rule::Fail3, 'baz', 123];
@@ -165,7 +165,7 @@ final class ListOfRuleTest extends ProcessingTestCase
 			$this->rule->processValue(
 				$value,
 				new MultiValueArgs(
-					new RuleRuntimeMeta(EfficientTestRule::class, new EmptyArgs()),
+					new RuleRuntimeMeta(PhasedTestRule::class, new EmptyArgs()),
 					null,
 					null,
 					false,